# api/swagger/swagger.json
{
  "info": {
    "title": "Tenant Service",
    "description": "This service is responsible for segmenting service configuration between distinct client applications",
    "version": "1.0.0"
  },
  "swagger": "2.0",
  "paths": {
...
    "/tenants/{id}": {
      "get": {
        "summary": "Get Tenant By Id",
        "description": "Get Tenant By Id",
        "operationId": "getTenant",
        "tags": [
          "tenants"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "description": "Id of Tenant",
            "type": "string",
            "in": "path",
            "required": true
          },
          {
            "name": "x-fast-pass",
            "description": "Bypass Auth",
            "type": "boolean",
            "in": "header",
            "require": false
          }
        ],
        "responses": {
          "200": {
            "description": "Tenant",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Tenant"
            }
          },
          "404": {
            "description": "Not found",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/tenants/_apiKey/{apiKey}": {
      "get": {
        "summary": "Get Tenant By Api Key",
        "description": "Get Tenant By Api Key",
        "operationId": "getTenantByApiKey",
        "tags": [
          "tenants"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "apiKey",
            "description": "Api Key of Tenant",
            "type": "string",
            "in": "path",
            "required": true
          },
          {
            "name": "x-fast-pass",
            "description": "Bypass Auth",
            "type": "boolean",
            "in": "header",
            "require": false
          }
        ],
        "responses": {
          "200": {
            "description": "Tenant",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Tenant"
            }
          },
          "404": {
            "description": "Error",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/tenants": {
      "get": {
        "summary": "Get Tenants",
        "description": "Get Page of Tenant(s)",
        "operationId": "getTenants",
        "tags": [
          "tenants"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "search",
            "description": "Text Search String",
            "type": "string",
            "in": "query",
            "required": false
          },
          {
            "name": "page",
            "description": "Page Number ( 0 index based )",
            "type": "integer",
            "format": "int64",
            "required": false
          },
          {
            "name": "size",
            "description": "Page Size - default 10",
            "type": "integer",
            "format": "int64",
            "required": false
          },
          {
            "name": "x-fast-pass",
            "description": "Bypass Auth",
            "type": "boolean",
            "in": "header",
            "require": false
          }
        ],
        "responses": {
          "200": {
            "description": "PageResponse",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Tenant"
            }
          }
        }
      },
      "post": {
        "summary": "Save Tenant",
        "description": "Save Tenant",
        "operationId": "saveTenant",
        "tags": [
          "tenants"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "tenant",
            "description": "tenant",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Tenant"
            },
            "in": "body",
            "required": true
          },
          {
            "name": "x-fast-pass",
            "description": "Bypass Auth",
            "type": "boolean",
            "in": "header",
            "require": false
          }
        ],
        "responses": {
          "201": {
            "description": "Tenant",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Tenant"
            }
          },
          "400": {
            "description": "Bad request",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "put": {
        "summary": "Update Tenant",
        "description": "Update Tenant",
        "operationId": "updateTenant",
        "tags": [
          "tenants"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "tenant",
            "description": "tenant",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Tenant"
            },
            "in": "body",
            "required": true
          },
          {
            "name": "x-fast-pass",
            "description": "Bypass Auth",
            "type": "boolean",
            "in": "header",
            "require": false
          }
        ],
        "responses": {
          "200": {
            "description": "Tenant",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Tenant"
            }
          },
          "400": {
            "description": "Bad request",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/users": {
      "post": {
        "summary": "Save User",
        "description": "Save User",
        "operationId": "saveUser",
        "tags": [
          "users"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "user",
            "description": "user",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/User"
            },
            "in": "body",
            "required": true
          },
          {
            "name": "x-fast-pass",
            "description": "Bypass Auth",
            "type": "boolean",
            "in": "header",
            "require": false
          }
        ],
        "responses": {
          "201": {
            "description": "User",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Bad request",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "put": {
        "summary": "Update User",
        "description": "Update User",
        "operationId": "updateUser",
        "tags": [
          "users"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "user",
            "description": "user",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/User"
            },
            "in": "body",
            "required": true
          },
          {
            "name": "x-fast-pass",
            "description": "Bypass Auth",
            "type": "boolean",
            "in": "header",
            "require": false
          }
        ],
        "responses": {
          "200": {
            "description": "User",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Bad request",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Get User By Id",
        "description": "Get User By Id",
        "operationId": "getUser",
        "tags": [
          "users"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "description": "Id of User",
            "type": "string",
            "in": "path",
            "required": true
          },
          {
            "name": "x-fast-pass",
            "description": "Bypass Auth",
            "type": "boolean",
            "in": "header",
            "require": false
          }
        ],
        "responses": {
          "200": {
            "description": "User",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "Not found",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Error": {
      "type": "object",
      "required": [
        "errorMessage"
      ],
      "properties": {
        "errorMessage": {
          "type": "string"
        },
        "violations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ConstraintViolation"
          }
        }
      }
    },
    "ConstraintViolation": {
      "type": "object",
      "required": [
        "param",
        "message"
      ],
      "properties": {
        "param": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "ServiceName": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
...
    "Tenant": {
      "type": "object",
      "required": [
        "name",
        "timestamp",
        "status",
        "apiKey",
        "apiSecret"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "apiKey": {
          "type": "string"
        },
        "apiSecret": {
          "type": "string"
        },
        "services": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ServiceName"
          }
        },
        "timestamp": {
          "type": "number",
          "format": "date"
        }
      }
    },
    "User": {
      "type": "object",
      "required": [
        "firstname",
        "lastname",
        "email",
        "phoneNumber",
        "username",
        "password",
        "accountId",
        "tenantId"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "locale": {
          "type": "string"
        },
        "firstname": {
          "type": "string"
        },
        "lastname": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "accountId": {
          "type": "string"
        },
        "tenantId": {
          "type": "string"
        }
      }
    },
...
  },
...
  "securityDefinitions": {},
...
}